requirements.txt
google-generativeai==0.3.2
pysrt==1.1.2
customtkinter==5.2.1
flask==3.0.2
